requirements.txt
alembic==1.5.2
astroid==2.4.2
attrs==20.3.0
autopep8==1.3.2
bidict==0.21.2
blinker==1.4
botocore==1.20.27
cachelib==0.1.1
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Session==0.3.2
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
greenlet==1.0.0
gunicorn==20.0.4
iniconfig==1.1.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
